+++ src/Crummy/Phlack/WebHook/CommandInterface.php
@@ -3,57 +3,65 @@
 namespace Crummy\Phlack\WebHook;
 
 use Crummy\Phlack\Common\Encodable;
+use Guzzle\Common\ToArrayInterface;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 
-interface CommandInterface extends Encodable
+/**
+ * CommandInterface
+ */
+interface CommandInterface extends Encodable, ToArrayInterface
 {
     /**
-     * @param OptionsResolverInterface $resolver
+     * @param $key
+     * @return mixed
      */
-    public function setDefaultOptions(OptionsResolverInterface $resolver);
+    public function get($key);
 
     /**
      * @return string
+     * @deprecated Will be removed in 0.5.0
      */
     public function getChannelName();
 
     /**
      * @return string
+     * @deprecated Will be removed in 0.5.0
      */
     public function getChannelId();
 
     /**
      * @return string
+     * @deprecated Will be removed in 0.5.0
      */
     public function getCommand();
 
     /**
      * @return string
+     * @deprecated Will be removed in 0.5.0
      */
     public function getTeamId();
 
     /**
      * @return string
+     * @deprecated Will be removed in 0.5.0
      */
     public function getText();
 
     /**
      * @return string
+     * @deprecated Will be removed in 0.5.0
      */
     public function getToken();
 
     /**
      * @return string
+     * @deprecated Will be removed in 0.5.0
      */
     public function getUserName();
 
     /**
      * @return string
+     * @deprecated Will be removed in 0.5.0
      */
     public function getUserId();
-
-    /**
-     * @return array
-     */
-    public function toArray();
 }